% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-types.R
\name{recode_multiple_choice}
\alias{recode_multiple_choice}
\title{Recode string responses to a multiple choice question.}
\usage{
recode_multiple_choice(frame, question_name, choices, ints = NULL)
}
\description{
Recode string responses to a multiple choice question.
}
